refactor: initial ts refactor
