+++ src/components/seo.tsx
@@ -1,13 +1,45 @@
-import React from "react"
+import React, { FunctionComponent } from "react"
 import { useStaticQuery, graphql } from "gatsby"
 import { useLocation } from "@reach/router"
 import { Helmet } from "react-helmet"
 
 import stripTrailingSlash from "../utils/strip-trailing-slash"
 
-const SEO = ({ title, description, image, article: isArticle, date, tags }) => {
+interface SEOProps {
+  title?: string
+  description?: string
+  image?: string
+  article?: boolean
+  date?: string
+  tags?: string[]
+}
+
+interface siteQueryData {
+  site: {
+    siteMetadata: {
+      defaultTitle: string
+      titleTemplate: string
+      defaultDescription: string
+      siteUrl: string
+      defaultImage: string
+      firstName: string
+      lastName: string
+      username: string
+      gender: string
+    }
+  }
+}
+
+const SEO: FunctionComponent<SEOProps> = ({
+  title,
+  description,
+  image,
+  article: isArticle,
+  date,
+  tags,
+}) => {
   const { pathname } = useLocation()
-  const { site } = useStaticQuery(
+  const { site }: siteQueryData = useStaticQuery(
     graphql`
       query SEO {
         site {
@@ -65,6 +97,7 @@ const SEO = ({ title, description, image, article: isArticle, date, tags }) => {
       <meta property="og:type" content={type} />
       {isArticle && <meta property="article:published_time" content={date} />}
       {isArticle &&
+        tags &&
         tags.map((tag) => <meta property="article:tag" content={tag} />)}
       {isArticle && (
         <meta property="article:author" content={stripTrailingSlash(siteUrl)} />
