--- src/utils/strip-trailing-slash.js
@@ -1,4 +1,0 @@
-const stripTrailingSlash = (string) =>
-  string.endsWith("/") ? string.slice(0, -1) : string
-
-export default stripTrailingSlash
