+++ src/utils/strip-trailing-slash.ts
@@ -1,0 +1,4 @@
+const stripTrailingSlash = (str: string): string =>
+  str.endsWith("/") ? str.slice(0, -1) : str
+
+export default stripTrailingSlash
